Metadata parsing - Fix inheritance and create more modules (#14)
* Tests - Metadata - Remove unused method parameter

The `fname` is used in the fixture and does not need passing to the test
methods anymore.

* Tests - Metadata - Refactor to parameterize the fixture

Change the test setup to parameterize the test fixture to enable class
level scoping. This now only parses the file once per case invocation.

* Tests - Metadata - Include column test case into main class

Refactor the test setup and include the test for expected columns with
the central test class.

* IO - Extract date handling to own module

Extract logic to parse dates to actual datetime object to its own
module. Helps with testing too, which are added with this commit.

* IO - MetaData - Use new DateManager module to handle dates.

Module was created in a previous commit and this removes all logic from
the MetaData class to use the module instead.

* Code Quality - reformat some lines for readability and typo fixes.

Also optimizes import statements in touched files.

* Tests - Move tests of same module into dedicated folder

Only affects the IO module

* IO - Start LocationManager module to parse lat/lon information

Extract logic to extract the lat/lon location information to its own
module. Tests will follow and most of this logic is passively covered by
sample file tests.

* Campaign - SnowEx - Rename variable mappings of YAML files

Rename snowex campaign specific YAML files and add a `snowex_` prefix to
better indicate and prevent future conflicts when more are added.

* IO - Locations - Refactor to rely on mapped keys based of YAML file

Refactor the location metadata logic to use the mapped codes from the
YAML file. This reduces identifying header fields to one location. If we
need to support more default headers, then I suggest to add them to the
base YAML file.

* IO - Introduce YamlCodes module to map defined codes as constants

Add module as a global registry for defined codes in metadata YAML to
reusable string constants in code.

* IO - Metadata - Refactor to rely on the YAML definitions

Refactor the Metadata class to use keys that were mapped with the YAML
files to detect header values. Also add TODOs for currently missing
fields in the YAML files.

* Tests - Start class structure tests for SnowEx campaign

* Profiles - Pep8 on base file

* BugFix - PointDataCollection - Move primary and metadata parser

The definition on which metadata and primary variable files to use needs
to be declared on the PROFILE_DATA_CLASs to allow overwriting the
definitions in inheriting classes.

* Run Pep8 on campaign.py

* Run Pep8 on test_dates.py

* Fix Flake8 CI raised errors.