Refactor - Only allow one overwrite file per variable/metadata
Major rework on how a user can specify overwrites for metadata or variables and
limit this information to one file per type. Also change the underlying
architecture and delegate the metadata parser class information to the profile.
This also reduces the number of arguments when parsing an actual file via
`read_csv` and delegates required information to the metadata class.

Still need to update the documentation once a first PR review iteration approves
the code changes.