IO - Locations - Fix return type indicator (#19)
* IO - Locations - Fix return type indicator

Fix the return type indicator for method 'parse_utm_epsg'. 'None' was
never returned and it is always an int

* GitHub CI - Add python 3.12

Personally I am developing on 3.12, so why not also add that to the
GitHub CI.

* IO - Locations - Raise error when no UTM info is found

Improve the `parse_utm_epsg` method by raising an error if no
information can be found instead of returning nothing at all.