Refactor - User YAML file handling and Metadata architecture (#20)
* IO - Locations - Fix return type indicator

Fix the return type indicator for method 'parse_utm_epsg'. 'None' was
never returned and it is always an int

* GitHub CI - Add python 3.12

Personally I am developing on 3.12, so why not also add that to the
GitHub CI.

* IO - Locations - Raise error when no UTM info is found

Improve the `parse_utm_epsg` method by raising an error if no
information can be found instead of returning nothing at all.

* MeasurementDescription - Fix `cast_type` definition, add validators

Fix the definition of the `cast_type` attribute to store the optional
value as a string. Also add type validation to the `code` and
`map_from` attributes

* ExtendedVariables - Add `allow_map_failures` attribute

Expand the ExtendedVariables object and indicate whether a mapping
attempt is allowed to fail with no match found.

* Tests - Variables - Add the first round of basic tests.

With the friendly help of JetBrains AI.

* ExtendedVariables - Keep track of parsed YAML files

Add an attribute to the ExtendedVariable object that will keep track of
parsed files that were used to create this instance.

* Tests - Create global fixture to generate YAML variable files

Create a fixture that allows for quickly generating YAML files with
variable definitions.

* IO - MetaParser - Create constants spell out variable names

Define constants for characters that are used to split or detect header
information. Also spell out some variable names for readability.

* Campaign - Move SnowEx source URLs to dedicated module

* Refactor - Only allow one overwrite file per variable/metadata

Major rework on how a user can specify overwrites for metadata or variables and
limit this information to one file per type. Also change the underlying
architecture and delegate the metadata parser class information to the profile.
This also reduces the number of arguments when parsing an actual file via
`read_csv` and delegates required information to the metadata class.

Still need to update the documentation once a first PR review iteration approves
the code changes.

* Address Flake8 CI issues

* Fix argument and return type hints for older Python versions

Trying to address issues raised by CI, where the type hints could not be parsed
by Python 3.9 and 3.10

* Setup.py - Add Python 3.12

Tests are passing on CI.

* Authors - Add myself

* README - Update docs and remove outdated links