Layer data header regression (#24)
* Dev - Add pytest-mock for easier mocking integration

* Metadata - Dates - Move more logic to module

Move more date and time logic to the now named DateTimeManager (from
DateManager). Also fixes a regression where checking for a non-existent dict key
needs to use get() to not error out.

* Dev requirements - Bump pytest and pytest-cov

Have been using this successfully local

* IO - Metadata - Location EPSG parser, allow no values

Some coordinates from the data are coming in as lat/lon with no EPSG code.
Allowing the parser of the metadata EPSG lookup to account for that case.

* IO - Metadata - Location - Fix type hinting

The indicated type was not correct.

* IO - Metadata - Location - Type hint fix for Python 3.9